Fix loop() in closures, float.floor/ceil return Float
- loop() (recur) now works inside closures called from loop bodies.
  The tail-call optimization path was bypassing the LoopRecur catch;
  the TailCall handler in the loop evaluator now properly intercepts
  LoopRecur from closure results.
- float.floor and float.ceil now return Float instead of Int.
  floor(3.7) returns 3.0, not 3. Consistent with float pipeline
  expectations. float.round still returns Int (intentional).

+++ src/typechecker.rs
@@ -1767,16 +1767,16 @@ impl TypeChecker {
             Box::new(Type::Int),
         )));
 
-        // float.ceil: (Float) -> Int
+        // float.ceil: (Float) -> Float
         env.define("float.ceil".into(), Scheme::mono(Type::Fun(
             vec![Type::Float],
-            Box::new(Type::Int),
+            Box::new(Type::Float),
         )));
 
-        // float.floor: (Float) -> Int
+        // float.floor: (Float) -> Float
         env.define("float.floor".into(), Scheme::mono(Type::Fun(
             vec![Type::Float],
-            Box::new(Type::Int),
+            Box::new(Type::Float),
         )));
 
         // float.abs: (Float) -> Float
